Refactor messager extracting
